+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Version 1.6.0 (not released yet)
 - Refactor sounds allocation to avoid frequent allocation and freeing of DirectSound buffers when channels pool is
   full - should improve performance
 - Add `kill_messages` command that allows to disable printing information about kills in the chatbox and the game console
+- Fix hud msg never disappearing in spectate mode
 - Other stability fixes
 
 Version 1.6.0-beta (released 2020-12-12)